Move some code for creating Boogie programs to BoogieFactory, and remove some wrapper of Util methods
